+++ Goalscorer/Models/Association.swift
@@ -18,42 +18,7 @@ class Association: Object {
     
     let competitions = LinkingObjects(fromType: Competition.self, property: "association")
 
-    var competitionsTemp: [Competition] = []
-    var playersTemp: [Player] = []
-
-    convenience init(name: String, regionCode: String, latitude: Double, longitude: Double, competitions: [Competition], players: [Player]) {
-        self.init()
-
-        self.name = name
-        self.regionCode = regionCode
-        self.latitude = latitude
-        self.longitude = longitude
-
-        self.competitionsTemp = competitions
-        self.playersTemp = players
-    }
-
-    override static func ignoredProperties() -> [String] {
-        return ["competitionsTemp", "playersTemp"]
-    }
-
     var coordinate: CLLocationCoordinate2D {
         return CLLocationCoordinate2D(latitude: latitude, longitude: longitude)
-    }
-}
-
-// TODO: モデルにUIKitかきたくない問題
-import UIKit
-import FlagKit
-
-extension Association {
-
-    var image: UIImage? {
-        switch regionCode {
-        case "CAF", "CAS", "CEU", "CNA", "COC", "CSA", "WW":
-            return UIImage(named: regionCode)
-        default:
-            return Flag(countryCode: regionCode)?.image(style: .roundedRect)
-        }
     }
 }
