+++ Goalscorer/Models/DataInitializer.swift
@@ -13,8 +13,6 @@ final class DataInitializer {
 
     func initData() {
         initRealm()
-
-        importOldFromUserDefaults()
     }
 }
 
@@ -32,22 +30,5 @@ private extension DataInitializer {
         } catch {
             print(error.localizedDescription)
         }
-    }
-
-    func importOldFromUserDefaults() {
-        UserDefaultsDAO().loadOldFavorites()
-            .map { plainObject in
-                let realmObject = FavoriteScorer()
-                let scorers = RealmDAO<Scorer>().findAll().filter("url == '\(plainObject.url)'")
-                realmObject.scorer = scorers.first!
-                realmObject.createdAt = plainObject.createdAt
-                realmObject.lastReadAt = plainObject.lastReadAt
-                realmObject.lastUpdatedAt = plainObject.lastUpdatedAt
-                return realmObject
-            }
-            .forEach {
-                RealmDAO<FavoriteScorer>().add(t: $0)
-            }
-        UserDefaultsDAO().saveOldFavorites([])
     }
 }
