+++ Goalscorer/Models/OverallScorer.swift
@@ -13,14 +13,4 @@ import RealmSwift
 class OverallScorer: Object {
     dynamic var url: String = ""
     dynamic var competition: Competition!
-
-    convenience init(url: String) {
-        self.init()
-
-        self.url = url
-    }
-
-    func setProperties(competition: Competition) {
-        self.competition = competition
-    }
 }
