+++ Goalscorer/Models/Player.swift
@@ -15,26 +15,4 @@ class Player: Object {
     dynamic var name: String = ""
     dynamic var association: Association!
     dynamic var order: Int = 999
-
-    convenience init(name: String, order: Int? = nil) {
-        self.init()
-
-        self.name = name
-        self.order = order ?? 999
-    }
-
-    func setProperties(association: Association) {
-        self.association = association
-        self.url = "https://en.wikipedia.org/wiki/\(path)#Career_statistics"
-    }
-}
-
-private extension Player {
-
-    private var path: String {
-        return name
-            .replacingOccurrences(of: " ", with: "_")
-            // `-` を置換
-            .addingPercentEncoding(withAllowedCharacters: CharacterSet.urlPathAllowed)!
-    }
 }
